Update SONATYPE publishing repository url (#47)
* Update SONATYPE publishing for new env vars

This updates the SONATYPE publish to look for new ENV vars,
SONATYPE_CENTRAL_USERNAME/SONATYPE_CENTRAL_PASSWORD which contains the
changes for the new sonatype publishing. New envs are used to help with
the transition.

https://central.sonatype.org/publish/publish-portal-ossrh-staging-api/#plugins-tested-for-compatibility

* Update SONATYPE publishing repository url

This updates the SONATYPE url for the new staging API that serves as a
shim. https://central.sonatype.org/publish/publish-portal-ossrh-staging-api/#getting-started-for-maven-api-like-plugins